refactor: Auth 서비스와 DTO 구조 리팩토링
- `AuthService` 관련 DTO 패키지 재구성: service.dto와 presentation.dto.response로 리팩토링
- `LoginResponse` -> `LoginInfoDto`로 변경 및 관련 메서드, 테스트 수정
- `AuthService.login` 및 `kakaoLogin` 반환 타입 수정
- import 경로 수정 및 불필요한 코드 제거
