+++ backend/fittoring/src/main/java/fittoring/application/auth/presentation/AuthController.java
@@ -8,11 +8,11 @@ import fittoring.application.auth.presentation.dto.request.SignUpRequest;
 import fittoring.application.auth.presentation.dto.request.ValidateDuplicateLoginIdRequest;
 import fittoring.application.auth.presentation.dto.request.VerificationCodeRequest;
 import fittoring.application.auth.presentation.dto.request.VerifyPhoneNumberRequest;
-import fittoring.application.auth.presentation.dto.response.AuthTokenDto;
-import fittoring.application.auth.presentation.dto.response.LoginResponse;
 import fittoring.application.auth.service.AuthService;
 import fittoring.application.auth.service.PhoneVerificationFacadeService;
 import fittoring.application.auth.service.PhoneVerificationService;
+import fittoring.application.auth.service.dto.AuthTokenDto;
+import fittoring.application.auth.service.dto.LoginInfoDto;
 import fittoring.application.exception.OauthLoginException;
 import fittoring.application.mentoring.presentation.dto.response.MemberLoginResponse;
 import fittoring.config.auth.AuthRequired;
@@ -53,9 +53,9 @@ public class AuthController {
     @PostMapping("/login")
     public ResponseEntity<MemberLoginResponse> login(@RequestBody @Valid SignInRequest request,
                                                      HttpServletResponse httpResponse) {
-        LoginResponse response = authService.login(request.loginId(), request.password());
-        CookieWriter.write(httpResponse, response.authTokenDto());
-        return ResponseEntity.ok(response.memberLoginResponse());
+        LoginInfoDto loginInfo = authService.login(request.loginId(), request.password());
+        CookieWriter.write(httpResponse, loginInfo.authTokenDto());
+        return ResponseEntity.ok(new MemberLoginResponse(loginInfo.memberId()));
     }
 
     @AuthRequired
@@ -117,9 +117,9 @@ public class AuthController {
             throw new OauthLoginException("카카오 로그인 에러 : " + error + " : " + errorDescription);
         }
 
-        LoginResponse loginResponse = authService.kakaoLogin(code);
-        AuthTokenDto authTokenDto = loginResponse.authTokenDto();
-        MemberLoginResponse memberLoginResponse = loginResponse.memberLoginResponse();
+        LoginInfoDto loginInfoDto = authService.kakaoLogin(code);
+        AuthTokenDto authTokenDto = loginInfoDto.authTokenDto();
+        MemberLoginResponse memberLoginResponse = new MemberLoginResponse(loginInfoDto.memberId());
 
         if (authTokenDto.isLoginSuccess()) {
             CookieWriter.write(httpResponse, authTokenDto);
