+++ backend/fittoring/src/main/java/fittoring/application/auth/presentation/dto/response/KakaoUserInfoResponse.java
@@ -1,0 +1,4 @@
+package fittoring.application.auth.presentation.dto.response;
+
+public record KakaoUserInfoResponse(Long id) {
+}
