--- backend/fittoring/src/main/java/fittoring/application/auth/presentation/dto/response/LoginResponse.java
@@ -1,9 +1,0 @@
-package fittoring.application.auth.presentation.dto.response;
-
-import fittoring.application.mentoring.presentation.dto.response.MemberLoginResponse;
-
-public record LoginResponse(
-        MemberLoginResponse memberLoginResponse,
-        AuthTokenDto authTokenDto
-) {
-}
