--- backend/fittoring/src/main/java/fittoring/application/auth/service/dto/KakaoUserInfoResponse.java
@@ -1,4 +1,0 @@
-package fittoring.application.auth.service.dto;
-
-public record KakaoUserInfoResponse(Long id) {
-}
