+++ backend/fittoring/src/main/java/fittoring/application/auth/service/dto/LoginInfoDto.java
@@ -1,0 +1,9 @@
+package fittoring.application.auth.service.dto;
+
+import fittoring.application.mentoring.presentation.dto.response.MemberLoginResponse;
+
+public record LoginInfoDto(
+        Long memberId,
+        AuthTokenDto authTokenDto
+) {
+}
