+++ backend/fittoring/src/test/java/fittoring/application/auth/service/AuthServiceTest.java
@@ -7,8 +7,8 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import fittoring.IntegrationTestSupport;
 import fittoring.application.FixtureUtil;
 import fittoring.application.auth.presentation.dto.request.SignUpRequest;
-import fittoring.application.auth.presentation.dto.response.AuthTokenDto;
-import fittoring.application.auth.presentation.dto.response.LoginResponse;
+import fittoring.application.auth.service.dto.AuthTokenDto;
+import fittoring.application.auth.service.dto.LoginInfoDto;
 import fittoring.application.auth.repository.RefreshTokenRepository;
 import fittoring.application.exception.DuplicateLoginIdException;
 import fittoring.application.exception.MisMatchPasswordException;
@@ -131,13 +131,13 @@ class AuthServiceTest extends IntegrationTestSupport {
         String rawPassword = "password";
 
         //when
-        LoginResponse actual = authService.login(loginId, rawPassword);
+        LoginInfoDto actual = authService.login(loginId, rawPassword);
 
         //then
         RefreshToken refreshToken = refreshTokenRepository.findByTokenValue(actual.authTokenDto().refreshToken())
                 .orElseThrow(null);
         SoftAssertions.assertSoftly(softly -> {
-                    assertThat(actual.memberLoginResponse().memberId()).isEqualTo(mentee.getId());
+                    assertThat(actual.memberId()).isEqualTo(mentee.getId());
                     assertThat(actual.authTokenDto().accessToken()).isNotNull();
                     assertThat(actual.authTokenDto().refreshToken()).isNotNull();
                     assertThat(refreshToken).isNotNull();
